fix(Revolut): changed amount type to long

+++ src/integration/agents/src/main/java/se/tink/backend/aggregation/agents/nxgen/serviceproviders/banks/revolut/fetcher/investment/entity/BalanceEntity.java
@@ -13,7 +13,7 @@ public class BalanceEntity {
     private String type;
 
     @JsonProperty("amount")
-    private int amount;
+    private long amount;
 
     @JsonProperty("currency")
     private String currency;
@@ -26,7 +26,7 @@ public class BalanceEntity {
         return type;
     }
 
-    public int getAmount() {
+    public long getAmount() {
         return amount;
     }
